Move common opmode code to RobotHardware
Moved most common opmode initialization routines 
to RobotHardware class.  This is intended to simplify 
opmodes derived from RobotHardware, and to 
eventually break the coupling between 
Executive.StateMachine and AutoOpmode

Note that the last source of dependence is 
imuUtilities, but that needs to be handled in a way 
that doesn't create lag in Manual, or cause errors when the IMU is absent, such as while testing vision stand 
alone.