Executive.StateMachine based on RobotHardware, not AutoOpmode
imuUtilities instance is now owned by RobotHardware, 
and therefore by all OpModes, rather than just being 
a part of AutoOpmode.  This means the 
Executive.StateMachine can be used in any OpMode 
based on RobotHardware, not just AutoOpmode.

IMU Utilities may need additional tweaking to make it 
efficient when unused, and to prevent errors when 
the RC Phone is not connected to the Rev Hub/IMU.