Added FakeHardware and AccelerationLimiter class
AutoDrive unitTest sim with FakeRobotHardware
It is important in the fake hardware to be aware
that fields of child objects hide, but do not replace
parent object fields.

AccelerationLimiter Class, doesn't seem helpful at
this time because it can cause the robot to
orbit waypoints, following constant acceleration
paths.

Some changes were made to Mecanum and
MecanumNavigation, but these should be
transparent. The changes are to intermediate
interfaces, and none of the mathematics should
be different.