AutoDrive with Acceleration Limiter - Running but not effective
Troubleshooting a series of bugs, some of which are 
with the code, and others seem to be with the way 
unit tests are getting compiled.

AccelerationLimiter is implemented on a new 
Mecanum.Command object 
(identical to the old Mecanum.Motion?)
So this must be inserted into AutoDrive where 
Mecanum.Wheels once was.

Another loose end is that the current "Command" 
needs to be normalized when the inputs are 
greater than 1, or for that matter, when the 
combination of the inputs requires wheel inputs 
greater than 1.