AutoDrive: Command Normalization and mecNav.delta
AutoDrive:
Normalizing the command, such that the sum of 
the magnitudes of vx, vy, av are less than or 
equal to 1.
Also, instead of using plain conversion from 
Mecanum.commandToWHeels(command), using a 
new mecanumNavigation.
deltaWheelsFromBodyRelativeMotion(command)
which may be more aware of the geometry.
Not sure, actually, but it DOES emulate the 
previous behavior better.