Fix: Don't declare time in FakeRobotHardware
FakeRobotHardware is a subclass of RobotHardware, 
which has a field 'double time.'  Within Java, a 
subclass cannot override parent fields the way it 
can override parent methods. Instead, parent fields 
are 'hidden,' but still exist and will still be accessed 
whenever a parent type reference is holding a child 
object.  
Consequently, it is better NOT to declare 
the field within the child, to prevent confusion about 
which field is being accessed.  Even safer would be 
to access fields through getter methods, which can 
be overridden in child classes.