Basically Functional with tests - needs Blue/Red location logic
System is now basically working.

However, It has no logic for location, RED vs BLUE, 
and then will need tuning. This could be as simple 
as switching between two different objects of the 
class SampledLocationsNormalized.

Logic could be added to find the height of the 
skystone row, or the angle of the camera.

Logic could be added to tune the sampleLocations 
based on the position of the robot and the 
field of view of the camera.