Add Vision setup from geometry, refactoring pipeline normalization
Data structures used for setting up the pipeline 
search regions using normalized values is being 
refactored.  The previous system repeated too 
much code, and was inflexible.  
The new system is based on normalizedValues,
normalizedPairs, and normalizedRectangles, 
which each pop up a few times.  This allows 
the pipeline to eventually be defined based on 
an array list of normalizedRenctangles to 
process, and lets the enclosing class handle 
determination of detections from the data and 
circumstances.