Waypoint index 5 fix/hack, stone C waypoints added
Adjusted index 5 to be adjacent to the outer wall, 
instead of through it.  The grab will still be a bit 
poor, but will at least not cause a collision.

Added grab_C and pickup_C in case a third stone 
is going to be moved.

Removed 7 inch offset for second and third 
skystone pickups.

Added hacky toggle for 3 point alignment with 
5th stone at 90 degree rotation.