States: Added configurable stone number in interactive init.
Whether dropping stones or placing on the 
foundation, a configurable number of stones can be 
set from 0 to 3.

Added pre-alignment stage to both AlignOuter 
and AlignInner States which get the y and theta 
of the robot corrected before strafing in the 
x direction to the alignment and eventually the 
parking location.